improve proj structure, add gitignore and readme
