+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "[python]": {
+    "editor.insertSpaces": true,
+    "editor.tabSize": 4
+  }
+}
