+++ readme.md
@@ -1,0 +1,11 @@
+# Informatics HS20 - Hospital
+
+### Unofficial Student Project
+
+Just in case you want to extend this. For now, this is just a friendly introduction 👋
+
+- Getting Started
+- Installation
+- Usage
+- Roadmap
+- etc.
